third_party/WebKit/LayoutTests/fast/html/import-node-into-quirks-document.html

<!-- quirks mode -->

<script src="../../resources/js-test.js"></script>

<style>
    .casesensitive { border: 1px solid red; }
</style>

<iframe id="standards-mode"></iframe>

<script>
description('Bug 257221: Moving an element from a standards mode document to a quirks mode one should handle the case difference of the id and class attributes');

var iframeDoc = document.getElementById('standards-mode').contentDocument;

iframeDoc.open();
iframeDoc.write('<!DOCTYPE html><div id="caseSensitive" class="caseSensitive">caseSensitive</div>');
iframeDoc.close();

div = iframeDoc.getElementById('caseSensitive');
document.body.appendChild(div);
shouldBe('document.querySelector("#caseSensitive")', 'div');
shouldBe('document.getElementById("caseSensitive")', 'div');
shouldBeEqualToString('getComputedStyle(div).borderTopWidth', '1px');
</script>
