third_party/blink/web_tests/platform/generic/fast/html/import-node-into-quirks-document-expected.txt

Bug 257221: Moving an element from a standards mode document to a quirks mode one should handle the case difference of the id and class attributes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.querySelector("#caseSensitive") is div
PASS document.getElementById("caseSensitive") is div
PASS getComputedStyle(div).borderTopWidth is "1px"
PASS successfullyParsed is true

TEST COMPLETE

caseSensitive
